[refactor] Refactor commands into subfolder layout
