+++ kicad_lib_manager/__init__.py
@@ -2,4 +2,4 @@
 KiCad Library Manager - Library management utilities for KiCad
 """
 
-__version__ = "0.2.2"
+__version__ = "0.2.3"
