+++ kicad_lib_manager/commands/add_3d/__init__.py
@@ -1,0 +1,3 @@
+from .command import add_3d
+
+__all__ = ["add_3d"]
