+++ kicad_lib_manager/commands/add_3d/command.py
@@ -7,8 +7,8 @@ from pathlib import Path
 
 import click
 
-from ..config import Config
-from ..utils.metadata import (
+from ...config import Config
+from ...utils.metadata import (
     CLOUD_METADATA_FILE,
     generate_env_var_name,
     get_default_cloud_metadata,
