+++ kicad_lib_manager/commands/add_hook/__init__.py
@@ -1,0 +1,3 @@
+from .command import add_hook
+
+__all__ = ["add_hook"]
