+++ kicad_lib_manager/commands/config/__init__.py
@@ -1,0 +1,3 @@
+from .command import config
+
+__all__ = ["config"]
