+++ kicad_lib_manager/commands/init/__init__.py
@@ -1,0 +1,3 @@
+from .command import init
+
+__all__ = ["init"]
