+++ kicad_lib_manager/commands/list_libraries/__init__.py
@@ -1,0 +1,3 @@
+from .command import list_cmd
+
+__all__ = ["list_cmd"]
