+++ kicad_lib_manager/commands/pin/__init__.py
@@ -1,0 +1,3 @@
+from .command import pin
+
+__all__ = ["pin"]
