+++ kicad_lib_manager/commands/pin/command.py
@@ -6,8 +6,8 @@ import sys
 
 import click
 
-from ..library_manager import find_kicad_config, list_libraries
-from ..utils.env_vars import (
+from ...library_manager import find_kicad_config, list_libraries
+from ...utils.env_vars import (
     expand_user_path,
     find_environment_variables,
     update_pinned_libraries,
