+++ kicad_lib_manager/commands/setup/__init__.py
@@ -1,0 +1,3 @@
+from .command import setup
+
+__all__ = ["setup"]
