+++ kicad_lib_manager/commands/status/__init__.py
@@ -1,0 +1,3 @@
+from .command import status
+
+__all__ = ["status"]
