+++ kicad_lib_manager/commands/template/__init__.py
@@ -1,0 +1,3 @@
+from .command import template
+
+__all__ = ["template"]
