+++ kicad_lib_manager/commands/unpin/__init__.py
@@ -1,0 +1,3 @@
+from .command import unpin
+
+__all__ = ["unpin"]
