+++ kicad_lib_manager/commands/update/__init__.py
@@ -1,0 +1,3 @@
+from .command import check_for_library_changes, update
+
+__all__ = ["check_for_library_changes", "update"]
